+++ src/plugins/intel_gpu/src/graph/graph_optimizer/prepare_primitive_fusing.cpp
@@ -275,7 +275,7 @@ void prepare_primitive_fusing::fuse_bias(program &p) {
             for (size_t i = 0; i < const_shape.size(); ++i) {
                 if (const_shape[i] != 1) {
                     count_elements_not_one++;
-                    idx_element_not_one = i;
+                    idx_element_not_one = static_cast<int32_t>(i);
                 }
                 if (count_elements_not_one > 1)
                     break;
