+++ src/plugins/intel_gpu/src/plugin/ops/custom.cpp
@@ -199,10 +199,10 @@ void CreateCustomOp(Program& p, const std::shared_ptr<ngraph::Node>& op, CustomL
             IE_THROW() << "Invalid input tensor for index: " << iidx;
         auto inputDims = op->get_input_shape(iidx);
 
-        xDim = inputDims[inputDims.size() - 1];
-        yDim = dims.size() > 1 ? inputDims[inputDims.size() - 2] : 0;
-        featureDim = dims.size() > 2 ? inputDims[inputDims.size() - 3] : 0;
-        batchDim = dims.size() > 3 ? inputDims[inputDims.size() - 4]: 0;
+        xDim = static_cast<int>(inputDims[inputDims.size() - 1]);
+        yDim = dims.size() > 1 ? static_cast<int>(inputDims[inputDims.size() - 2]) : 0;
+        featureDim = dims.size() > 2 ? static_cast<int>(inputDims[inputDims.size() - 3]) : 0;
+        batchDim = dims.size() > 3 ? static_cast<int>(inputDims[inputDims.size() - 4]) : 0;
     }
     const std::map<char, int> vars = {
         { 'b', batchDim }  , { 'B', batchDim },
